Reordering files so they are more organize
